scripts/requirements.txt
moviepy
librosa
soundfile
matplotlib
pillow
opencv-python
numpy
scikit-learn
svgpathtools